Revert "Companion package (#1)"
This reverts commit 7e8cd38a68c4113b0a575993260140f039ad29c6.
