+++ .github/workflows/build.yml
@@ -22,7 +22,7 @@ jobs:
         uses: jupyterlab/maintainer-tools/.github/actions/base-setup@v1
 
       - name: Install dependencies
-        run: python -m pip install -U "jupyterlab==4.2.6"
+        run: python -m pip install -U "jupyterlab>=4.0.0,<5"
 
       - name: Lint the extension
         run: |
@@ -33,12 +33,11 @@ jobs:
       - name: Build the extension
         run: |
           set -eux
-          jlpm
-          jlpm build
-          pip install -e ".[test]"
-          jupyter labextension develop . --overwrite
+          python -m pip install .[test]
 
+          jupyter labextension list
           jupyter labextension list 2>&1 | grep -ie "jupyter-iframe-commands.*OK"
+          python -m jupyterlab.browser_check
 
       - name: Package the extension
         run: |
@@ -157,32 +156,16 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-
       - name: Setup Python
         uses: actions/setup-python@v5
         with:
           python-version: '3.12'
-
-      - name: Base Setup
-        uses: jupyterlab/maintainer-tools/.github/actions/base-setup@v1
-
-      - name: Download extension package
-        uses: actions/download-artifact@v4
-        with:
-          name: extension-artifacts
-
       - name: Install the dependencies
         run: |
-          set -eux
           python -m pip install -r requirements.txt
-          jlpm
-          jlpm build
-          python -m pip install .
-
       - name: Build the JupyterLite site
         run: |
           jupyter lite build --contents README.md --output-dir ghpages/lite
-
       - name: Upload artifact
         id: deployment
         uses: actions/upload-pages-artifact@v3
@@ -190,7 +173,7 @@ jobs:
           path: ./ghpages
 
   deploy:
-    needs: build-lite
+    needs: build
     if: github.ref == 'refs/heads/main'
     permissions:
       pages: write
