+++ .github/workflows/check-release.yml
@@ -1,9 +1,9 @@
 name: Check Release
 on:
   push:
-    branches: ['main']
+    branches: ["main"]
   pull_request:
-    branches: ['*']
+    branches: ["*"]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
@@ -15,21 +15,12 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-
       - name: Base Setup
         uses: jupyterlab/maintainer-tools/.github/actions/base-setup@v1
-
-      - name: Install the dependencies
-        run: |
-          set -eux
-          python -m pip install -r requirements.txt
-          jlpm
-          jlpm build
-          python -m pip install .
-
       - name: Check Release
         uses: jupyter-server/jupyter_releaser/.github/actions/check-release@v2
         with:
+
           token: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Upload Distributions
