+++ .gitignore
@@ -127,6 +127,3 @@ dmypy.json
 # lite build for demo
 ghpages/lite
 .jupyterlite.doit.db
-
-# nx cache files
-.nx
