+++ .prettierignore
@@ -2,6 +2,5 @@ node_modules
 **/node_modules
 **/lib
 **/package.json
+!/package.json
 jupyter_iframe_commands
-
-/.nx/workspace-data
