+++ jupyter_iframe_commands/__init__.py
@@ -5,10 +5,12 @@ except ImportError:
     # in editable mode with pip. It is highly recommended to install
     # the package from a stable release or in editable mode: https://pip.pypa.io/en/stable/topics/local-project-installs/#editable-installs
     import warnings
-
     warnings.warn("Importing 'jupyter_iframe_commands' outside a proper installation.")
     __version__ = "dev"
 
 
 def _jupyter_labextension_paths():
-    return [{"src": "labextension", "dest": "jupyter-iframe-commands"}]
+    return [{
+        "src": "labextension",
+        "dest": "jupyter-iframe-commands"
+    }]
