--- lerna.json
@@ -1,5 +1,0 @@
-{
-  "$schema": "node_modules/lerna/schemas/lerna-schema.json",
-  "version": "independent",
-  "npmClient": "yarn"
-}
