--- nx.json
@@ -1,15 +1,0 @@
-{
-  "namedInputs": {
-    "default": ["{projectRoot}/**/*"]
-  },
-  "targetDefaults": {
-    "build": {
-      "cache": true,
-      "dependsOn": ["^build"],
-      "outputs": [
-        "{projectRoot}/packages/extension/lib",
-        "{projectRoot}/packages/host/lib"
-      ]
-    }
-  }
-}
