+++ pyproject.toml
@@ -40,7 +40,6 @@ exclude = [".github", "binder"]
 "jupyter_iframe_commands/labextension" = "share/jupyter/labextensions/jupyter-iframe-commands"
 "install.json" = "share/jupyter/labextensions/jupyter-iframe-commands/install.json"
 
-
 [tool.hatch.build.hooks.version]
 path = "jupyter_iframe_commands/_version.py"
 
@@ -68,7 +67,7 @@ version_cmd = "hatch version"
 
 [tool.jupyter-releaser.hooks]
 before-build-npm = [
-    "python -m pip install 'jupyterlab==4.2.6'",
+    "python -m pip install 'jupyterlab>=4.0.0,<5'",
     "jlpm",
     "jlpm build:prod"
 ]
