+++ schema/plugin.json
