+++ src/index.ts
@@ -10,7 +10,7 @@ import { ISettingRegistry } from '@jupyterlab/settingregistry';
 
 import { expose, windowEndpoint } from 'comlink';
 
-import { CommandBridge } from 'jupyter-iframe-commands-host';
+import { CommandBridge } from './api';
 
 /**
  * A plugin to expose an API for interacting with JupyterLab from a parent page.
@@ -45,7 +45,7 @@ const plugin: JupyterFrontEndPlugin<void> = {
     }
 
     const { commands } = app;
-    const api = new CommandBridge('jupyterlab', { commands });
+    const api = new CommandBridge({ commands });
     const endpoint = windowEndpoint(self.parent);
     expose(api, endpoint);
   }
