+++ tsconfig.json
@@ -13,8 +13,11 @@
     "noUnusedLocals": true,
     "preserveWatchOutput": true,
     "resolveJsonModule": true,
+    "outDir": "lib",
+    "rootDir": "src",
     "strict": true,
     "strictNullChecks": true,
     "target": "ES2018"
-  }
+  },
+  "include": ["src/*"]
 }
